requirements.txt
anyio==4.12.1
appnope==0.1.4
asttokens==3.0.0
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
comm==0.2.3
contourpy==1.3.3
curl_cffi==0.13.0
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
executing==2.2.1
Flask==3.1.3
fonttools==4.60.1
frozendict==2.4.7
geopandas==1.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.1.7
multitasking==0.0.12
nest-asyncio==1.6.0
numpy==2.3.4
packaging==25.0
pandas==2.3.3
parso==0.8.5
peewee==4.0.2
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.4.0
pluggy==1.6.0
prompt_toolkit==3.0.52
protobuf==7.34.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==3.0
Pygments==2.19.2
pyogrio==0.11.1
pyparsing==3.2.5
pyproj==3.7.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-telegram-bot==22.6
pytz==2025.2
pyzmq==27.0.2
qrcode==8.2
requests==2.33.0
ruff==0.15.7
scipy==1.17.0
seaborn==0.13.2
setuptools==82.0.1
shapely==2.1.2
six==1.17.0
soupsieve==2.8.3
stack-data==0.6.3
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.3
wcwidth==0.2.13
websockets==16.0
Werkzeug==3.1.6
wheel==0.46.3
xgboost==3.1.3
yfinance==1.2.0
supabase==2.0.1